feat: 컴포넌트의 size가져올 수 있는 패턴과 커스텀 Hook 추가
