+++ src/Frontend/utils/remToPixel.ts
@@ -1,7 +1,6 @@
 import { theme } from '@Styles/theme';
 
 const remToPixel = (remValue: string) : number => {
-  console.log("remvalue in rem to pixel : ", remValue);
   const remNum = parseInt(remValue.substr(0, remValue.length - 3));
   return remNum * theme.mainFontSize;
 }
